Corrected computeFrameSnapshot to fetch the container using a helper method (which correctly accounts for joint frames), which required moving it to storeMethods.
